+++ homebanking/src/main/java/com/mindhub/homebanking/services/TransactionService.java
@@ -4,5 +4,5 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 
 public interface TransactionService {
-   ResponseEntity<Object> performTransaction(Integer amount, String description, String sourceAccountNumber, String destinationAccountNumber, Authentication authentication);
+   ResponseEntity<Object> performTransaction(Double amount, String description, String sourceAccountNumber, String destinationAccountNumber, Authentication authentication);
 }
